condensed the code
