--- src/constants.js
@@ -1,6 +1,0 @@
-const CONSTANTS = {
-	MODULE_NAME: "ignore-vision",
-	PATH: `modules/ignore-vision/`
-};
-CONSTANTS.PATH = `modules/${CONSTANTS.MODULE_NAME}/`;
-export default CONSTANTS;
